requirements.txt
appdirs==1.4.3
cachetools==2.0.0
certifi==2017.4.17
chardet==3.0.3
dill==0.2.6
Django==1.10.7
enum34==1.1.6
future==0.16.0
gapic-google-cloud-vision-v1==0.90.3
google-auth==1.0.1
google-auth-httplib2==0.0.2
google-cloud-core==0.24.1
google-cloud-vision==0.24.0
google-gax==0.15.13
googleapis-common-protos==1.5.2
grpcio==1.3.5
httplib2==0.10.3
idna==2.5
oauth2client==3.0.0
olefile==0.44
packaging==16.8
Pillow==4.1.1
ply==3.8
proto-google-cloud-vision-v1==0.90.3
protobuf==3.3.0
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.2.0
pytesseract==0.1.6
requests==2.17.3
rsa==3.4.2
six==1.10.0
urllib3==1.21.1
whitenoise==3.3.0
